requirements.txt
requests==2.32.3
beautifulsoup4==4.12.3
pdfplumber==0.11.4
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
pydantic==2.10.3
loguru==0.7.3
python-dotenv==1.0.1
lxml==5.3.0
fastapi==0.115.0
uvicorn==0.30.6
PyJWT==2.12.0
argon2-cffi==23.1.0
email-validator==2.2.0
httpx==0.27.2
